--- src/hooks/useToast/useToast.ts
@@ -1,10 +1,0 @@
-import { useContext } from "react";
-import { ToastContext } from "../../context/toastContext";
-
-export const useToast = () => {
-  const context = useContext(ToastContext);
-  if (!context) {
-    throw new Error("useToast는 ToastProvider 안에서만 사용할 수 있습니다.");
-  }
-  return context;
-};
